out_0402_modify: 修改无线传输的短ID为4字节

+++ code/src/stack/custom/mac/mac.h
@@ -11,7 +11,7 @@
 #ifndef __MAC_H__
 #define __MAC_H__
 
-extern uint16_t mac_short_addr;
+extern uint32_t mac_short_addr;
 extern uint64_t mac_long_addr;
 
 extern uint8_t mac_seq;
